authentication-poc/requirements.txt
httpx==0.24.1
python-dotenv==1.0.0
PyJWT==2.8.0
tenacity==8.2.3
pytest==7.4.3